fix : camera fix as a single script
