--- Assets/Scripts/CameraFocusedFollowingUnit.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f8346a6c7544c5949a6becc461b0d0a7
