--- Assets/Scripts/CameraFocusedSelectedUnit.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 442129d7548eb6c4395d5e7e974d5de8
